+++ service/favorite_service.go
@@ -2,7 +2,7 @@ package service
 
 import "douyin/repository"
 
-func CreateFavorite(user int64, feed int64) bool {
+func CreateFavorite(user uint, feed uint) bool {
 	favorite := repository.Favorite{User: user, Feed: feed, Cancel: 0}
 	err := repository.NewFavoriteDaoInstance().CreateFavorite(&favorite)
 	if err != nil {
@@ -11,7 +11,7 @@ func CreateFavorite(user int64, feed int64) bool {
 	return true
 }
 
-func DeleteFavorite(user int64, feed int64) bool {
+func DeleteFavorite(user uint, feed uint) bool {
 	err := repository.NewFavoriteDaoInstance().DeleteFavorite(user, feed)
 	if err != nil {
 		return false
